+++ data/src/main/java/com/microsoft/azure/kusto/data/auth/PublicAppTokenProviderBase.java
@@ -28,6 +28,13 @@ public abstract class PublicAppTokenProviderBase extends MsalTokenProviderBase {
         super(clusterUrl, authorityId, httpClient);
     }
 
+    // This is synchronized as we don't want users to be prompt many times on async operations.
+    // This way the first thread gets here, finishes and token can be retrieved silently by other threads.
+    @Override
+    protected synchronized String acquireAccessTokenImpl() throws DataServiceException, DataClientException {
+        return super.acquireAccessTokenImpl();
+    }
+
     @Override
     protected void initializeWithCloudInfo(CloudInfo cloudInfo) throws DataClientException, DataServiceException {
         super.initializeWithCloudInfo(cloudInfo);