+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Automatic retries for queued ingestion
 * `executeQuery`, `executeMgmt` to call with a specific type.
+### Fixed 
+* Timer was used if authentication throws after client was closed 
+* Public client credentials (user prompt,device auth) are synchronized - so that users are prompt once
 
 ## [5.0.0]
 ### Fixed